backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
cryptography==41.0.7
websockets==12.0
python-socketio==5.11.0
aiohttp==3.9.1
httpx[http2]==0.26.0
openai==1.10.0
anthropic==0.18.1
email-validator==2.1.0
numpy>=1.24.0

# Keycloak OIDC Integration (Production-grade)
cachetools>=5.3.0
tenacity>=8.2.0
